trade_assistant/requirements.txt
telethon
pandas
python-binance
python-dotenv